Remove copy-pasted unused variable
`loggedInAsField` is used by the Silverstripe Framework's MemberLoginForm
which has no commonality with the SAMLLoginForm. They share a common
ancestor of LoginForm, but are MemberLoginForm is not also an ancestor
of SAMLLoginForm.

As such the single consumption place for the property in the
MemberLoginForm is never utilised by SAMLAuthenticator.

It seems like it was probably a copy-paste & modify job that just didn't
identify the property as unnecessary.